pyproject.toml
[project]
name = "zenoh-ci"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eclipse-zenoh>=1.6.2",
    "flatbuffers>=25.9.23",
    "numpy>=2.3.4",
    "zenoh-cli>=0.6.8",
]
